requirements.txt
Django==5.2.7
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
django-cors-headers==4.9.0
PyMySQL==1.1.2
Pillow==11.3.0
reportlab==4.4.4
channels==4.0.0
channels-redis==4.2.0
redis==5.0.1
pytz==2025.2
django-filter==25.2